Add ability to use separator to EngNumber
This now gives the user the opportunity to use a separator between the
number part and the magnitude suffix, so I could write `1.2 k` for
instance, if I passed a space as a separator.

The default is an empty string, so users not using this feature will be
unaffected by this change.

I also added the facility to EngUnit, which calls EngNumber in the
implementation.